FormatOWL update
- replaced RDF with OWL in enumeration of SchemaDoc.cs
- FormatOWL defined as subclass of IFormatExtension
- first version of FormatOWL::FormatDefinitions to generate the computer
interpretable listings
- addition of methods FormatOWL::FormatEnumerationCompListing and
FormatOWL::FormatSelectCompListing to implement the official conversion
pattern that is used to generate the OWL file. This conversion pattern
is currently not shown in the html documentation because poorly
informative.
- modified arguments of IFormatExtension::FormatSelect (added map and
included). The same for its subclasses.
- modification in DocumentationISO.cs to pass the variable
"view_included" instead of "included" to the function FormatDefinitions.
- preliminary working generation of a subset of OWL files in AnnexA
(different versions for each model view)
- initial draft of FormatOWL::FormatEntity